clean up all compiler warnings in the bzl build

+++ src/aig/gia/giaRex.c
@@ -430,7 +430,7 @@ void Gia_ManAutomWalkOne( Gia_Man_t * p, int nSteps, Vec_Wrd_t * vStates, Vec_In
                 continue;
             Index = Vec_WrdFind( vStates, pNext[k] );
             Count = Index == -1 ? 0 : Vec_IntEntry( vCounts, Index );
-            if ( CountMin > Count || (CountMin != ABC_INFINITY && Count && ((float)CountMin / Count) > (float)rand()/RAND_MAX ) )
+            if ( CountMin > Count || (CountMin != ABC_INFINITY && Count && ((double)CountMin / Count) > (double)rand()/RAND_MAX ) )
             {
                 CountMin = Count;
                 IndexMin = Index;
